more parts are working for the menu
logins are created now for hr and admin and employee there are 2 supporting csv files necessary ftEmployeeData.csv and ptEmployeeData.csv ft stores full time emps and pt stores part time emps as they have different data fields they can not be saved in the same csv the menu class contains a csv reader that reads both files and places data in a hash map using the name of emps as the key (loadEmployees). there is a writer that saves all of the employees and their data back on to the csvs everytime an emp is created and when the final quit on the first menu it saves(saveEmployees). run method contains code for the actual menu constructor creates hash maps for both types of emps. also inits a scanner for the system, and calls load employees to load all the data into the hash maps. the class payrollsim is used to run the run method in payrollsystem, and creates a new instance of the menu. Let me if you find ways to break it the last step will be adding error prevention. to make it unbreakable as they will try to break the code.